requirements.txt
torch
langchain_community
python-dotenv
streamlit
fastapi
uvicorn
requests
langchain_chain
faiss-cpu
bs4